split solver data tests
